fix: behebt zirkul\u00e4ren Import f\u00fcr TextParser
